@@ -1,28 +1,33 @@
 // #![allow(unused_imports, deprecated, unused_must_use, unused_mut, unused_variables, dead_code, unreachable_code)]
 
-#[cfg(test)]
-#[macro_use]
-extern crate futures_await_test;
-
 pub use {
     cherry::Cherry,
-    connection::PoolConfig,
-    datasource::DataSource,
-    types::Transaction,
+    adapter::AnyArguments,
+    adapter::AnyRow,
+    anyhow::Error,
+    // connection::PoolConfig,
+    // datasource::DataSource,
+    // types::Transaction,
 };
 
 pub(crate) mod cherry;
-pub(crate) mod connection;
-pub(crate) mod query;
-pub(crate) mod datasource;
+mod adapter;
+// pub(crate) mod connection;
+// pub(crate) mod query;
+// pub(crate) mod datasource;
 
-pub mod types;
+// pub mod types;
 
-pub mod error {
-    pub use anyhow::Error;
-}
+// pub mod anyhow {
+//     pub use anyhow::*;
+// }
 
 pub mod sqlx {
+    pub use sqlx::*;
+}
+
+
+/*pub mod sqlx {
     pub use sqlx::{Arguments, Database, Decode, Encode, Row, types::Type};
     #[cfg(feature = "json")]
     pub use sqlx::types::Json;
@@ -37,7 +42,7 @@ pub mod sqlx {
     // pub use sqlx::sqlite::{Sqlite, SqliteArguments, SqliteQueryResult, SqliteRow};
     // #[cfg(feature = "mssql")]
     // pub use sqlx::mssql::{Mssql, MssqlArguments, MssqlQueryResult, MssqlRow};
-}
+}*/
 
 #[cfg(not(any(feature = "mysql", feature = "postgres", feature = "sqlite", feature = "mssql")))]
 compile_error!("one of the features ['mysql', 'postgres', 'sqlite', 'mssql'] must be enabled");
